conform to asp.net core extension conventions
