+++ samples/Routine.Samples.Basic/Startup.cs
@@ -11,7 +11,7 @@ namespace Routine.Samples.Basic
     {
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddRoutineDependencies();
+            services.AddRoutine();
         }
 
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
