+++ src/Routine/BuildRoutine.cs
@@ -46,54 +46,6 @@ namespace Routine
 
     public static class BuildRoutineExtensions
     {
-        #region AspNetCore
-
-        public static IServiceCollection AddRoutineDependencies(this IServiceCollection source) => source.AddRoutineDependencies<JsonSerializerAdapter>();
-        public static IServiceCollection AddRoutineDependencies<TJsonSerializer>(this IServiceCollection source) where TJsonSerializer : class, IJsonSerializer =>
-            source
-                .AddHttpContextAccessor()
-                .AddSingleton<IJsonSerializer, TJsonSerializer>();
-
-        public static IApplicationBuilder UseRoutine(this IApplicationBuilder source,
-            Func<CodingStyleBuilder, ICodingStyle> codingStyle,
-            Func<ServiceConfigurationBuilder, IServiceConfiguration> serviceConfiguration = null,
-            IRestClient restClient = null,
-            IJsonSerializer serializer = null,
-            ICache cache = null,
-            Func<InterceptionConfigurationBuilder, IInterceptionConfiguration> interceptionConfiguration = null
-        )
-        {
-            serviceConfiguration ??= s => s.FromBasic();
-            interceptionConfiguration ??= i => i.FromBasic();
-
-            return source.UseMiddleware<RoutineMiddleware>(
-                BuildRoutine.Context()
-                    .Using(
-                        restClient: restClient,
-                        serializer: serializer,
-                        cache: cache,
-                        interceptionConfiguration: interceptionConfiguration(BuildRoutine.InterceptionConfig())
-                    )
-                    .AsServiceApplication(serviceConfiguration, codingStyle)
-            );
-        }
-
-        public static IApplicationBuilder UseRoutineInDevelopmentMode(this IApplicationBuilder source)
-        {
-            ReflectionOptimizer.Disable();
-
-            return source;
-        }
-
-        public static IApplicationBuilder UseRoutineInProductionMode(this IApplicationBuilder source)
-        {
-            ReflectionOptimizer.Enable();
-
-            return source;
-        }
-
-        #endregion
-
         #region ContextBuilder
 
         public static IClientContext AsClientApplication(
